CATTY-595 continue refactoring class
