+++ src/Catty/ViewController/Download/Details/ProjectDetailStoreViewControllerNewReportExtension.swift
@@ -22,8 +22,8 @@
 
 import Foundation
 
-@objc
-extension ProjectDetailStoreViewController2 {
+
+extension ProjectDetailStoreViewController {
     func reportProject() {
         guard UserDefaults.standard.bool(forKey: NetworkDefines.kUserIsLoggedIn) else {
             Util.alert(text: kLocalizedLoginToReport)
